+++ app/src/test/java/com/zegoggles/smssync/AppTest.java
@@ -15,7 +15,7 @@ public class AppTest {
     }
 
     @Test public void shouldGetVersionCode() throws Exception {
-        assertThat(App.getVersionCode(RuntimeEnvironment.getApplication())).isEqualTo(1806);
+        assertThat(App.getVersionCode(RuntimeEnvironment.getApplication())).isEqualTo(1808);
     }
 
     @Test public void shouldTestOnSDCARD() throws Exception {